requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
Flask==3.1.2
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
requests==2.32.5
urllib3>=1.24.2,<2.4.0
Werkzeug==3.1.3
zipp==3.23.0

# Database dependencies
SQLAlchemy>=2.0.29
PyMySQL==1.1.0
cryptography==41.0.7
